requirements.txt
alembic==0.9.6
amqp==2.2.2
aniso8601==1.3.0
APScheduler==3.4.0
asyncio==3.4.3
Babel==2.5.1
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.3
bitarray==0.8.1
bleach==2.1.2
blinker==1.4
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
Cheetah==2.4.4
click==6.7
colorama==0.3.9
docopt==0.6.2
document==1.0
docx==0.2.4
dominate==2.3.1
elasticsearch==6.0.0
enum34==1.1.6
Flask==0.12.2
Flask-APScheduler==1.7.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.3
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-MySQLdb==0.2.0
Flask-PageDown==0.2.2
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
flower==0.9.2
fn==0.4.3
ForgeryPy==0.1
funcsigs==1.0.2
futures==3.2.0
hdfs==2.1.0
html5lib==1.0.1
httpie==0.9.9
httpie-negotiate==1.0.0
idna==2.6
impyla==0.14.0
itsdangerous==0.24
Jinja2==2.10
kafka==1.3.5
kombu==4.1.0
lxml==4.1.1
Mako==1.0.7
Markdown==2.6.10
MarkupSafe==1.0
mysql-connector-python-rf==2.2.2
MySQL-python==1.2.5
mysqlclient==1.3.7
nltk==3.2.5
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pyasn1==0.4.2
pyDes==2.0.1
Pygments==2.2.0
pyhs2==0.6.0
pykerberos==1.1.14
pypinyin==0.28.0
python-dateutil==2.6.1
python-docx==0.8.6
python-editor==1.0.3
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-kerberos==0.11.0
rsa==3.4.2
sasl==0.2.1
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.15
suds==0.4
thrift==0.10.0
thrift-sasl==0.3.0
tornado==4.5.2
typing==3.6.2
tzlocal==1.5.1
urllib3==1.22
vine==1.1.4
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.13
WTForms==2.1
xlrd==1.1.0
